requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
cloudinary==1.40.0
dj-database-url==2.2.0
Django==5.0.6
django-appconf==1.0.6
django-ckeditor-5==0.2.13
django-cloudinary-storage==0.3.0
django-js-asset==2.2.0
django_compressor==4.5
idna==3.7
pillow==10.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.6.0
